requirements.txt
# aws-cdk-lib
requests
constructs
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.core
aws-cdk.aws-iam
awscli